Improve telemetry to not requre any buffers
This is achieved through the use of the chunks namespace. The types defined allow for "virtually concatinating" in memory buffers, removing the need for memcpys into a single buffer before encoding a packet. This commit also includes a test suite to ensure that both the chunking and COBS is up to specification